+++ recipes/install_php.rb
@@ -26,8 +26,7 @@ node.override['php']['conf_dir'] = '/etc/php/7.2'
 
 include_recipe "php::package"
 
-node['php']['module_packages'].each do |package_name, do_install|
-  next unless do_install
+Ingenerator::Php::PhpModules.module_packages_to_install(node).each do | package_name |
   package package_name
 end