+++ spec/recipes/install_php_spec.rb
@@ -1,7 +1,8 @@
 require 'spec_helper'
 
 describe 'ingenerator-php::install_php' do
-  let (:chef_run) { ChefSpec::SoloRunner.new.converge 'ingenerator-php::install_php' }
+  let (:chef_runner) { ChefSpec::SoloRunner.new }
+  let (:chef_run)    { chef_runner.converge 'ingenerator-php::install_php' }
 
   it "installs php from package with the community cookbook recipe" do
     expect(chef_run).to include_recipe "php::package"
@@ -49,19 +50,17 @@ describe 'ingenerator-php::install_php' do
     end
   end
 
-  it "installs the php-apc module from package" do
-    expect(chef_run).to install_package "php-apc"
-  end
-
-  it "installs the php5-curl module from package" do
-    expect(chef_run).to install_package "php5-curl"
+  %w(php-apcu php-apcu-bc php7.2-curl).each do | pkg |
+    it "installs the #{pkg} module from package" do
+      expect(chef_run).to install_package pkg
+    end
   end
 
   context "with optional modules in node attributes" do
       let (:chef_run) do
         ChefSpec::SoloRunner.new do |node|
           node.normal['php']['module_packages']['php-gd'] = true
-          node.normal['php']['module_packages']['php-apc'] = false
+          node.normal['php']['module_packages']['php-apcu'] = false
         end.converge(described_recipe)
       end
 
@@ -70,8 +69,29 @@ describe 'ingenerator-php::install_php' do
       end
 
       it "does not install packages for modules that are disabled" do
-        expect(chef_run).not_to install_package 'php-apc'
+        expect(chef_run).not_to install_package 'php-apcu'
       end
   end
+
+  context 'with invalid legacy packages in node attributes' do
+
+    it 'throws if requesting any package with a php5 name' do
+      chef_runner.node.default['php']['module_packages']['php5-imagick'] = true
+      expect { chef_run }.to raise_exception /php5 package php5-imagick/
+    end
+
+    it 'throws if requesting php-apc name' do
+      chef_runner.node.default['php']['module_packages']['php-apc'] = true
+      expect { chef_run }.to raise_exception /php5 package php-apc/
+    end
+
+    it 'throws if requesting multiple bad packages' do
+      chef_runner.node.default['php']['module_packages']['php5-any'] = true
+      chef_runner.node.default['php']['module_packages']['php5-other'] = true
+      expect { chef_run }.to raise_exception /php5 package php5-any, php5-other/
+    end
+
+  end
+
 
 end