+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/). Note that
 
 ## Unreleased
 
+* [BREAKING] Update default module packages to their php7 versions and throw if a 
+  project requests any php5 module packages
 * [BREAKING] Update share_env_inis paths to /etc/php/7.2/... per ubuntu package
   locations - will throw if the project still specifies any at /etc/php5
 * [BREAKING] Expect to install php7.2